+++ CHANGELOG.md
@@ -2,8 +2,25 @@
 
 All notable changes to this project will be documented in this file.
 
-## v1.0.7
+## Unreleased
 
-- UMD, improved gulpfile.js. Refactor test folder.
+Log of unreleased changes.
 
-- riccioStore, getInstance(), isInitialized(). Deprecated getStore().
+### Added
+
+- UMD support.
+- Added `getInstance()`, `isInitialized()` functions.
+- Added `buildLayout()` function. Called on init and when a media query is
+triggered.
+
+### Changed
+
+- Improved Gulp tasks.
+- Refactor test folder.
+
+### Deprecated
+
+- Deprecated `getStore()` function. To get an initialized Riccio instance use
+`getInstance()`.
+- Deprecated `Riccio.prototype.needs()`, `Riccio.prototype.setRows()` and
+`Riccio.prototype.setItems()` in favor of `Riccio.prototype.buildLayout()`.